Separated logic from game file.
