+++ rock.json
@@ -1,10 +1,10 @@
 {
   "name": "Timmy",
-  "mood": "Bored",
-  "hunger": 6,
-  "boredom": 3,
-  "energy": 4,
-  "canFeed": true,
+  "mood": "Sad",
+  "hunger": 5,
+  "boredom": 7,
+  "energy": 3,
+  "canFeed": false,
   "canPlay": true,
   "zeroEnergyTurns": 0
 }
