+++ rock/src/main/java/com/petrock/PetRockController.java
@@ -1,18 +1,35 @@
 package com.petrock;
 
+import java.io.File;
+import java.io.FileReader;
 import java.util.Scanner;
+
+import com.google.gson.Gson;
 
 public class PetRockController {
     private PetRock rock;
     private PetRockView view;
     
-    public PetRockController(PetRock rock, PetRockView view) {
+    public PetRockController(PetRockView view) {
         this.rock = rock;
         this.view = view;
     }
     
     public void startGame() {
         Scanner scanner = new Scanner(System.in);
+        PetRock rock;
+        
+        File file = new File("rock.json");
+        if (file.exists()) {
+            try (FileReader reader = new FileReader(file)) {
+                Gson gson = new Gson();
+                rock = gson.fromJson(reader, PetRock.class);
+            } catch (Exception e) {
+                rock = createNewRock(scanner);
+            }
+        } else {
+            rock = createNewRock(scanner);
+        }
         while (true) {
             System.out.println("Choose an action:");
             System.out.println("1. Feed the Rock");
@@ -29,10 +46,20 @@ public class PetRockController {
                     rock.saveToFile();
                     System.out.println("Goodbye!");
                     System.exit(0);
+                    scanner.close();
                 }
                 default -> System.out.println("Invalid choice. Try again.");
             }
             view.displayStatus(rock);
         }
+  
     }
+    private static PetRock createNewRock(Scanner scanner) {
+        System.out.print("Enter a name for your pet rock: ");
+        String name = scanner.nextLine();
+        return new PetRock(name);
+    }
+
+
+
 }
